backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==1.1.0
spacy==3.7.2
openai==0.28.1
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.0.3
faiss-cpu==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
pydantic==2.5.0
azure-storage-blob==12.19.0
azure-identity==1.15.0
# vapi-python==0.1.0  # Not needed - using HTTP requests directly
